renamed html-routes to handlebars-routes server line 38: required handlebars-routes instead of html routes
