+++ server.js
@@ -35,7 +35,7 @@ app.set("view engine", "handlebars");
 // Routes
 // =============================================================
 require("./routes/api-routes.js")(app);
-require("./routes/html-routes.js")(app);
+require("./routes/handlebars-routes.js")(app);
 
 db.sequelize.sync({ force: true }).then(function () {
     app.listen(PORT, () => console.log(`Listening on port ${PORT}...`));
